requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pypdf==4.3.1
python-docx==1.1.2
scikit-learn==1.5.2
transformers==4.44.2
torch>=2.1.0
